LayoutTests/svg/animations/getCurrentTime-pause-unpause-setCurrentTime-expected.txt


PASS SVGSVGElement.getCurrentTime() with setCurrentTime() when animation is paused and unpaused
